sync2/requirements.txt
adal==1.2.3
azure-common==1.1.25
azure-servicebus==0.50.2
bson==0.5.9
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
DateTime==4.3
dnspython==1.16.0
idna==2.9
isodate==0.6.0
msrest==0.6.14
msrestazure==0.6.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tzlocal==2.1
uamqp==1.2.7
urllib3==1.25.9
zope.interface==5.1.0
